requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg==3.2.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
redis==6.2.0
sqlparse==0.5.3
typing_extensions==4.14.0
